Extracted ElementsPanel from Components.js
