+++ src/js/shot-generator-new/components/InspectedElement/HandInspector/HandPresetsEditor/index.js
@@ -73,8 +73,7 @@ React.memo(({
 
   updateObject,
   createHandPosePreset,
-  withState,
-  scene
+  withState
 }) => {
   const thumbnailRenderer = useRef()
 
@@ -149,7 +148,7 @@ React.memo(({
       })
       let skeleton = sceneObject.skeleton
       let model = sceneObject.model
-      let originalSkeleton = scene.children.filter(child => child.userData.id === id)[0].getObjectByProperty("type", "SkinnedMesh").skeleton.bones
+      let originalSkeleton = Object.values(skeleton)
       let handSkeleton = {}
       setSelectedHand(handName)
       for(let i = 0; i < originalSkeleton.length; i++) {
